requirements.txt
colorama==0.4.6
extras==1.0.0
Flask==3.1.1
flask_limiter==3.12
Flask_Login==0.6.3
flask_wtf==1.2.2
Markdown==3.4.1
Pillow==11.2.1
psutil==5.9.0
psycopg2==2.9.10
rapidfuzz==3.13.0
Requests==2.32.4
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
